requirements.txt
numpy>=1.21.0
pandas>=1.3.0
gym>=0.21.0
torch>=1.9.0
stable-baselines3>=1.5.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
seaborn>=0.11.2
tqdm>=4.62.3
cupy-cuda11x>=10.0.0  # Adjust based on your CUDA version
plotly>=5.3.0
psutil>=5.8.0
python-dateutil>=2.8.2
pytest>=6.2.5
pytest-cov>=2.12.0
# PostgreSQL support
psycopg2-binary>=2.9.0
# Docker support (if running in containers)